requirements.txt
pypdf2
pillow
reportlab
paddleocr
paddlepaddle
